[mxfp8 moe training] use dim1 cast cuda kernel in bwd
stack-info: PR: https://github.com/pytorch/ao/pull/2897, branch: danielvegamyhre/stack/64
